trying to complete the timescaleengine ( state is not working yet )
git-svn-id: svn://svn.code.sf.net/p/qwt/code/trunk@1591 29e17f8d-cf37-471e-a288-72cc03a76eda

+++ qwt/playground/timescale/timeinterval.h
@@ -15,11 +15,11 @@ public:
 
     TimeInterval rounded( TimeDate::IntervalType ) const;
 
-    TimeInterval adjusted( int stepSize,
+    TimeInterval adjusted( double stepSize,
         TimeDate::IntervalType ) const;
 
-    int width( TimeDate::IntervalType ) const;
-    int roundedWidth( TimeDate::IntervalType ) const;
+    double width( TimeDate::IntervalType ) const;
+    double roundedWidth( TimeDate::IntervalType ) const;
 
 private:
     QDateTime d_minDate;
